trying to get Station Summary Components to render
